requirements.txt
customtkinter>=5.2.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0